refactor: 대소문자 구분 없이 enum 사용 가능하도록 수정
