+++ src/main/java/com/project/undefined/company/entity/Company.java
@@ -34,14 +34,6 @@ public class Company extends BaseEntity {
     @Enumerated(value = EnumType.STRING)
     private Region region;
 
-    public enum Series {
-        A, B, C, D, E, F, IPO
-    }
-
-    public enum Region {
-        SEOUL, ETC
-    }
-
     public static Company of(final String name, final Series series, final Region region) {
         return new Company(null, name, series, region);
     }
