+++ src/test/java/com/project/undefined/acceptance/company/CompanyTest.java
@@ -10,8 +10,8 @@ import com.project.undefined.acceptance.utils.RestAssuredUtils;
 import com.project.undefined.common.dto.response.ErrorResponse;
 import com.project.undefined.company.dto.request.CreateCompanyRequest;
 import com.project.undefined.company.entity.Company;
-import com.project.undefined.company.entity.Company.Region;
-import com.project.undefined.company.entity.Company.Series;
+import com.project.undefined.company.entity.Region;
+import com.project.undefined.company.entity.Series;
 import com.project.undefined.company.repository.CompanyRepository;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
